First commit for Cassandane.
This will be a Perl-based integration test suite for Cyrus.

The initial commit is just a rewrite of the genmail.pl script which I
used earlier to create test data for the "pop3showafter" annotation.

Why "Cassandane" ? Wikipedia indicates that Cassandane was the name of
the consort of King Cyrus the Great of Persia, founder of the Achaemenid
Persian Empire.  So that's kinda cool.