Add IMAPMessageStore::get_server_name()
which uses a feature in the latest Mail::IMAPTalk to extract the IMAP
server's connection banner and to parse it for the Cyrus server name.
This is useful when checking that we have connected to the server we
wanted to connect to, which is particulatly important when we talk to
two servers at once, e.g. for replication testing.