Be more precise when parsing RFC822 headers.
The regexp for parsing an RFC822 header name now uses the actual
definition of a header name from RFC2822, not some approximate rubbish I
made up on the spot.  Discovered when an LKML archive failed to parse
due to someone's mailer adding a header with an underscore in it's name.
The lesson is: always test with real data too.