Make instance names more unique.
If two cyrus instances were started up in the same second, they would
end up using the same pathname for the root directory.  This is fine as
long as you only ever want one instance running, but for replication
tests we need two separate instances.  Work around this by creating the
default names of instances using the timestamp of the creation of the
first instance plus a counter.