Fix wandering working directory
due to the supposedly scope-local chdir in _start_master() actually
leaking the changed working directory into the main Cassandane process
thanks to a missing "local".

Changed Instance::_start_master() to use the Instance::run_utility()
code path for fork+exec instead of using the Perl system().  This gives
us one place to put all the UNIX things that need to be done when
spawning new processes to run the Code Under Test, whether it be master
or reconstruct or anything else.

Put the chdir() and setrlimit() in that code path, in the child process
between fork() and exec().  This also means we don't need the
File::chdir module, the builtin Perl chdir() function is perfectly
adequate.