Disconnect during replication
The test runs "sync_client -u" to do a replication run.  The defined and
correct behaviour of that code is to exit early with an exit code of 1
if any of the user's folders cannot be synced because they are locked.
Unfortunately the test has open an IMAP session to each of the master
and the replica, so the chances of the mailbox being locked are high.
Work around this by disconnected and reconnecting around the replication
run.

Also, don't assert on the return value of run_utility, as run_utility
does that checking now and more usefully.