Add annotation handling to Message.
The Message object can now handle getting, setting and listing
annotations.  These are treated as special kinds of attributes, i.e.
metadata for the message which is not visible in the canonical string
form of the message.  Indeed, get_annotation() is a wrapper around
get_attribute() and set_annotation() likewise.

Also, the c'tor is extended to parse ANNOTATION data as reported by the
IMAP FETCH command and converted by Mail::IMAPTalk.  This allows
annotations to be transparently set on Messages read from an
IMAPMessageStore if the correct magic is used with
IMAPMessageStore::set_fetch_attributes().

These changes allow Cassandane tests to check the values of per-message
annotations using the same check_messages() paradigm already used.