Pull replication support into C::C::TestCase
and out of C::Instance.  This way it's much easier to run because it's
better integrated.  Also we only need to start exactly the correct
number of instances for each test (albeit through the horrible hack of
automatically starting a replica if the test name contains the word
'replication' which it just so happens is already the case).