Shut down the annotator daemon gracefully
or indeed, at all.  The Annotator tests were hanging if they failed, due
apparently to annotator.pl backgrounding itself (despite being told not
to), and not being shut down.  So allow it to go background, read the
pidfile it writes anyway, and use the existing pidfile handling code
designed for the master process to shut down the annotator.