Better gdb integration for Cyrus services
Services that run from the Cyrus master are difficult to debug at the
best of times, and sometimes you really need to get them running under
gdb.  The existing mode='gdb' feature of Instance::_fork_utility() was
just able to do that, as it required the binary to be run with a
controlling terminal.

So replace mode='gdb' with a new one, using the existing undocumented -D
option to all Cyrus services, a small trampoline program which helps a
human run gdb at the right time with the right arguments, and
cassandane.ini to decide which services get the gdb treatment.