More sensible logging of Instance startup
Move the creation of Instances and the associated logging from the
Cassandane::Cyrus::TestCase ctor into a new method which is called from
the set_up method.  This avoids spamming the logs with many unhelpful
"basedir" messages when Cassandane is run with a test filter (because
the filter feature of the Test::Unit infrastructure creates TestCase
instances for all tests and those that fail the filter are thrown away
again).

Also, make the "basedir" messages more helpful by printing the name of
the test and the role the instance plays in the test.