Provide a better way of tweaking config per-test
If a method called config_FOO exists for test test_FOO, then it's called
with a Cassandane::Config instance that it can futz with, just before
any Instances are created in the test set_up.  This is just as weird and
hacky as the restart_with_config method previously used for the one test
where this is necessary, but avoids creating unused instances, which
both reduces log spam and makes it easier to attach a gdb to the correct
imapd.