ThreadedGenerator sets up uid & cid attributes
of the messages it generates.

UIDs are generated using a simple counter, which can be initialised or
changed using the new method set_next_uid(); default is suitable for
filling an empty mailbox.

CIDs are tracked per-thread and calculated using the new
Message::make_cid() method, which was moved from a sub in
C::C::Conversations, as it's really a property of (some) messages.