Added keyed_on parameter to check_messages
By default, messages in the actual and expected hashes are keyed on
message subject, but this is inconvienent when the test deliberately
generates messages with the same subject (e.g. because they're identical
copies).  So add a 'keyed_on' parameter which is the name of a Message
accessor to use to get the key for the hashes.  'uid' is a useful value.