Emit only a single result summary in TAP mode
with a single count of tests and a single list of failures, instead of
one for each .pm file.  This is far less confusing.  As a nice side
effect, all the complicated code for handling scheduling of suites and
tests by name has now moved from testrunner.pl to a new class
Cassandane::Unit::TestPlan.