Kill master processes harder
by sending signals SIGQUIT, SIGILL and SIGKILL in order.  This should
ensure that we clean up master regardless, and should get a core if the
graceful shutdown doesn't work, e.g. because of BZ3616.  Also ensure the
enclosing test fails with a useful message if sending anything stronger
than SIGQUIT is necessary in $instance->stop().