Keep recording results after 1st parallel failure
due to the Perl &&= operator not evaluating the RHS if the LHS is
already false, thus incorrectly skipping the call to _finish_workitem()
after the first failure or error is reported.  Use bitwise &= instead,
but make sure we don't have undef on the RHS.

"You keep using that operator.  I do not think it means what you think
it means."