Add Cassini->bool_val()
which interprets the value from cassandane.ini as a boolean.
'yes', 'true' 'on' and '1' all map to true (case insensitive);
'no', 'false' 'off' and '0' all map to false.  Other string die.

There's only once place where this is used at the moment, but I'm about
to add a second one.