Unified xlog() and process stderr/stdout
Each test captures stdout & stderr in a temporary log file, and adds it
to the test's annotation strings so that it's printed should the test
fail.  xlog() already emits to stderr as well as syslog.  Several of the
more interesting Cyrus processes will (sometimes) initialise syslog with
LOG_PERROR which emits to stderr.  The result is that almost all the
interesting errors are captured together, and emitted only on error,
even if -v is specified.  Hopefully this will make it easier to diagnose
failing tests.