Parse cyrus.conf of pre-existing on-disk instances
So that we have Service objects which reflect the pre-existing
instance's port assignments, so that re-starting a pre-existing instance
will wait for the correct ports to be opened by master.

The "start-instance.pl --latest" command was failing to start correctly
if port 9100 was not allocated to a service in the existing instance.