cyr_buildinfo: Read Cyrus build information at instance creation.
Summary:
This adds the buildinfo output of the cyr_buildinfo tool to
the Cassandane instance. It's main use is to decide which tests to
run based on Cyrus compilation settings.

Make sure you run Cyrus with https://git.cyrus.foundation/D88

Test Plan: See the SearchFuzzy test suite for an example.

Reviewers: #imap_reviewers, rost

Reviewed By: rost

Differential Revision: https://git.cyrus.foundation/D93