explicitly expect Cassandane modules to be in '.'
Newer perls don't automatically include '.' in @INC for security
reasons.

We expect our Cassandane modules to be in the current directory,
so make that explicit.

Modules not in our distribution are included before the use lib
line, to ensure we use the installed version, not some rogue
version in cwd.