TestCase: :needs_foo_bar feature tests can now take a (value)
If your test function just needs the "foo.bar" feature to be enabled
(i.e. cyr_buildinfo reports it as "true"), then attribute it with
:needs_foo_bar as usual

If your test function needs the foo.bar feature to have a particular
value, then attribute it with :needs_foo_bar(value)

Real-world example incoming in a moment!